Domainlerde Ekleme ve düzeltmeler yapıldı

+++ Core/Project.Domain/Entities/Concretes/Product.cs
@@ -15,7 +15,7 @@ namespace Project.Domain.Entities.Concretes
         public bool CanBeProduced { get; set; }
         public bool IsReadyMade { get; set; }
 
-        public int Quantity { get; set; }
+        public decimal Quantity { get; set; }
 
         public decimal UnitPrice { get; set; }
         public int UnitId { get; set; }
